fix: Move logging to its own class
