+++ lib/osml/model_endpoint/roles/me_http_role.ts
@@ -88,8 +88,6 @@ export class MEHTTPRole extends Construct {
     )!;
 
     // Defining constants for better readability
-    const ME_CONTAINER_REPOSITORY_NAME =
-      this.meContainerConfig.ME_CONTAINER_REPOSITORY;
     const MR_FIRELENS_LOG_GROUP_NAME = `/aws/${this.mrDataplaneConfig.METRICS_NAMESPACE}/MRFireLens`;
     const MR_SERVICE_LOG_GROUP_NAME = `/aws/${this.mrDataplaneConfig.METRICS_NAMESPACE}/MRService`;
     const MR_HTTPENDPOINT_LOG_GROUP_NAME = `/aws/${this.mrDataplaneConfig.METRICS_NAMESPACE}/HTTPEndpoint`;
@@ -149,7 +147,7 @@ export class MEHTTPRole extends Construct {
         "ecr:DescribeRepositories"
       ],
       resources: [
-        `arn:${this.partition}:ecr:${props.account.region}:${props.account.id}:repository/${ME_CONTAINER_REPOSITORY_NAME}`
+        `arn:${this.partition}:ecr:${props.account.region}:${props.account.id}:repository/*`
       ]
     });
